Refactor tools module and add function tool conversion utility
- Rename `tools.py` to `tools.py` in `openai_shared` for better organization
- Add `function_tool_to_responses_tool_param` to convert
  `agents.FunctionTool` to `FunctionToolParam`
- Introduce import for `FunctionToolParam` to support new functionality
